--- CodeCampusHUBT/NCKH2024TQQD/src/main/java/com/codecampushubt/NCKH2024TQQD/Nckh2024TqqdApplication.java
@@ -1,13 +1,0 @@
-package com.codecampushubt.NCKH2024TQQD;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Nckh2024TqqdApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Nckh2024TqqdApplication.class, args);
-	}
-
-}
