--- CodeCampusHUBT/NCKH2024TQQD/src/main/resources/application.properties
@@ -1,16 +1,0 @@
-spring.application.name=NCKH2024TQQD
-server.port=3000
-
-
-#
-#   config data source data-jpa
-#
-spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=HubtCampus;encrypt=true;trustServerCertificate=true
-spring.datasource.username=sa
-spring.datasource.password=1
-
-#
-# Tranh truong hop JPA tu dong doi ten truong tu camelCase sang snake_case
-#
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-
