--- CodeCampusHUBT/NCKH2024TQQD/src/test/java/com/codecampushubt/NCKH2024TQQD/Nckh2024TqqdApplicationTests.java
@@ -1,13 +1,0 @@
-package com.codecampushubt.NCKH2024TQQD;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Nckh2024TqqdApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
