+++ NCKH2024TQQD/.gitignore
@@ -31,3 +31,19 @@ build/
 
 ### VS Code ###
 .vscode/
+
+# Không commit file cấu hình chứa thông tin nhạy cảm
+application.properties
+application.yml
+application-dev.properties
+application-prod.properties
+.env
+
+# Nếu dùng Docker hoặc Kubernetes
+docker-compose.override.yml
+secrets/
+
+# Nếu có chứa thông tin nhạy cảm trong logs
+logs/
+*.log
+
